% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-methods.R
\name{coerce,SummarizedExperiment,data.table-method}
\alias{coerce,SummarizedExperiment,data.table-method}
\title{Coerce a SummarizedExperiment object to a long data.table, retaining the data in all assays, rowData, colData}
\usage{
\S4method{coerce}{SummarizedExperiment,data.table}(from, to)
}
\arguments{
\item{from}{[`SummarizedExperiment`]}

\item{to}{[`character`]}
}
\value{
['data.table]
}
\description{
Coerce a SummarizedExperiment object to a long data.table, retaining the data in all assays, rowData, colData
}
