% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMetaAnalysisToDT.R
\name{convertMetaAnalysisToDT}
\alias{convertMetaAnalysisToDT}
\title{Parse gene signature metaanalysis .rds files into a data.table.}
\usage{
convertMetaAnalysisToDT(dataDir, saveDir, fileName = "gene_compound.csv")
}
\arguments{
\item{dataDir}{`character(1)` Path to the top level metaanalysis directory. Folders in
this directory should be named to match '*AnalyticPancan', where '*' is the molecular
data type of that signature file.}

\item{saveDir}{`character(1)` Path to write the output .csv to.}
}
\description{
Parse gene signature metaanalysis .rds files into a data.table.
}
