% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToCsv-methods.R
\name{.convertSEToDataTableList}
\alias{.convertSEToDataTableList}
\title{Convert the data in a SummarizedExperiment into a `list` of `data.table`s}
\usage{
.convertSEToDataTableList(SE)
}
\arguments{
\item{SE}{A [`SummarizedExperiment`] whos data will be converted to a `list`
of `data.table`s named for their source within the object.}
}
\description{
Convert the data in a SummarizedExperiment into a `list` of `data.table`s
}
\section{Note}{

  Any data stored in `metadata(SummarizedExperiment)`
  will will be converted to an ASCII representation
}

\keyword{internal}
