% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToCsv-methods.R
\name{.sensSlotToDataTables}
\alias{.sensSlotToDataTables}
\title{Convert the sensitivity slot of a PSet into a list of data.tables}
\usage{
.sensSlotToDataTables(sensSlot)
}
\arguments{
\item{sensSlot}{[`list`] The sensitivity slot of a PSet, as returned by `sensitivitySlot`}
}
\description{
Use this function to make writing data in the sensitivity slot to disk easy.
}
\keyword{internal}
