% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToCsv-methods.R
\name{.sensSlotToLong}
\alias{.sensSlotToLong}
\title{Merge all the data in the `sensitivity` slot into a single long format `data.table`}
\usage{
.sensSlotToLong(sensSlot)
}
\arguments{
\item{sensSlot}{[`list`] PharmacoSet sensitivity slot, as returned by `sensitivitySlot`.}
}
\value{
[`data.table`] Long format of data in sensitivity slot
}
\description{
Merge all the data in the `sensitivity` slot into a single long format `data.table`
}
