% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToCsv-methods.R
\name{.writeAnnotationToTxt}
\alias{.writeAnnotationToTxt}
\title{Write the annotation list in a PharmacoSet to a text file}
\usage{
.writeAnnotationToTxt(
  annotations,
  dsType,
  filePath,
  objectName,
  sepChar = "|||"
)
}
\arguments{
\item{annotations}{`list` Annotations as returned by the `annotaiton` function}

\item{dsType}{`character(1)`}

\item{filePath}{`character(1)` The path to save the output file to.}

\item{objectName}{`character(1)` The name of the PSet the annotations are from.}

\item{sepChar}{`character(1)` Separator to use when pasting together
multiple lines of a list item.}
}
\value{
Writes to disk, does not return.
}
\description{
Each list item is one line, lines in each item are separated with `sepChar` so they
 can be split into to their original stucture.
}
\keyword{internal}
