% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToCsv-methods.R
\name{.writeMolecularProfilesToCsv}
\alias{.writeMolecularProfilesToCsv}
\title{Convert each SummarizedExperiment in a CSet into a list of data.tables, then save to disk as .csv files}
\usage{
.writeMolecularProfilesToCsv(SElist, filePath, objectName)
}
\arguments{
\item{SElist}{A [`list`] of `SummarizedExperiment` objects, as returned by `molecularProfilesSlot`}

\item{filePath}{[`character`] Path to save the files .csv files}

\item{objectName}{[`character`] The name of the `CSet` object being written to disk}
}
\description{
Convert each SummarizedExperiment in a CSet into a list of data.tables, then save to disk as .csv files
}
