% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToParquet-methods.R
\name{.writeMolecularProfilesToParquet}
\alias{.writeMolecularProfilesToParquet}
\title{Convert each SummarizedExperiment in a CSet into a list of data.tables, then save to disk as .csv files}
\usage{
.writeMolecularProfilesToParquet(SElist, filePath, objectName, ...)
}
\arguments{
\item{SElist}{A [`list`] of `SummarizedExperiment` objects, as returned by `molecularProfilesSlot`}

\item{filePath}{[`character`] Path to save the files .csv files}

\item{objectName}{[`character`] The name of the `CSet` object being written to disk}

\item{...}{`pairlist` Fall through arguments to `arrow::write_parquet`.}
}
\description{
Convert each SummarizedExperiment in a CSet into a list of data.tables, then save to disk as .csv files
}
