% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToCsv-methods.R
\name{.writeSensitivityToCsv}
\alias{.writeSensitivityToCsv}
\title{Convert a PSets' sensitivty slot into a long data.table and save to disk as a .csv}
\usage{
.writeSensitivityToCsv(sensSlot, filePath, objectName)
}
\arguments{
\item{sensSlot}{[`list`] PharmacoSet sensitivity slot, as returned by `sensitivitySlot`.}
}
\description{
Convert a PSets' sensitivty slot into a long data.table and save to disk as a .csv
}
\keyword{internal}
