% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToParquet-methods.R
\name{.writeSensitivityToParquet}
\alias{.writeSensitivityToParquet}
\title{Convert a PSets' sensitivty slot into a long data.table and save to disk as a .csv}
\usage{
.writeSensitivityToParquet(sensSlot, filePath, objectName, ...)
}
\arguments{
\item{sensSlot}{[`list`] PharmacoSet sensitivity slot, as returned by `sensitivitySlot`.}
}
\description{
Convert a PSets' sensitivty slot into a long data.table and save to disk as a .csv
}
\keyword{internal}
