% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneAnnotationToCSV.R
\name{geneAnnotationToCSV}
\alias{geneAnnotationToCSV}
\title{Dump gene annotation files to .csv for use in PharmacoDI}
\usage{
geneAnnotationToCSV(
  filePath = file.path("../..", "data", "metadata"),
  pattern = "Gencode.v33.annotation.*"
)
}
\arguments{
\item{filePath}{[`character`] vector indiciating path to the annotation directory}

\item{pattern}{[`character`] files to match using `pattern` argument in `list.files` function}
}
\value{
NULL Writes files to `filePath`
}
\description{
Dump gene annotation files to .csv for use in PharmacoDI
}
