% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPSetGeneSigs.R
\name{getPSetGeneSigs}
\alias{getPSetGeneSigs}
\title{Read in all the gene signatures for a PharmacoSet and assemble into a single `data.table`}
\usage{
getPSetGeneSigs(filePath, pattern, save = TRUE)
}
\arguments{
\item{filePath}{[`character`] Where to search for the `pattern` argument to get the file paths for the
PSet gene signatures.}

\item{pattern}{[`character`] Passed to `list.files` to match all .rds files containing `pattern`. This
should usually be the PSet name as it will also be used as the file name if `save` is `TRUE`.}

\item{save}{[`logical`] Should the `data.table` be saved to disk as a .csv file? Default is `TRUE`.}
}
\value{
[`data.table`] Optionally returns the gene signature data.table for the specified PSet
}
\description{
Read in all the gene signatures for a PharmacoSet and assemble into a single `data.table`
}
