% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGeneSignatureToDT.R
\name{mergePSetGeneSigsToDT}
\alias{mergePSetGeneSigsToDT}
\title{Merge the results data for a list of `PharmacoSig` objects containing different gene signatures into a single,
   fully annotated long `data.table` and optionally save to disk as a .csv file}
\usage{
mergePSetGeneSigsToDT(geneSigL, saveDir, fileName, BPPARAM = bpparam())
}
\arguments{
\item{geneSigL}{[`list`] of `PharmacoSig` objects, as returned by `readGeneSigsForPSet`}

\item{saveDir}{Optional [`character`] vector specifying the path at which to save a .csv of the `data.table`}

\item{fileName}{Optional [`character`] vector specifying the name for the .csv file. If you specify one of `saveDir`
or `fileName`, you must specify both.}
}
\value{
A long [`data.table`] containing the gene signature statistics for each drug x gene x tissue combination
}
\description{
Merge the results data for a list of `PharmacoSig` objects containing different gene signatures into a single,
   fully annotated long `data.table` and optionally save to disk as a .csv file
}
