% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGeneSignatureFiles.R
\name{processGeneSignatureFiles}
\alias{processGeneSignatureFiles}
\title{Read gene signature files for many experiments and merge them to a 
  `data.table`}
\usage{
processGeneSignatureFiles(
  filePath,
  mDataType = c("rna", "cnv", "mutation"),
  ...,
  keyColumns = c("gene", "compound", "tissue", "dataset", "mDataType"),
  fileFormats = c("rds", "qs"),
  logDir = "logs"
)
}
\arguments{
\item{filePath}{`character(1)` Path to top level signature directory. This
should contain a folder `<mDataType>`.}

\item{mDataType}{`character(1)` The molecular data type of the signatures.}

\item{...}{Fall through parameters to [`BiocParallel::bplapply`]. These
parameters contral internal function parallelization.}

\item{keyColumns}{`character` Vector of column names in the signature file
to join on. For developer use, do no change.}

\item{fileFormats}{`character` Vector of file formats without a dot 
(e.g., rds)}
}
\value{
`data.table` The gene signatures properly formatted to a single 
  long format table.
}
\description{
Read gene signature files for many experiments and merge them to a 
  `data.table`
}
