% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGeneSignatureFiles.R
\name{processPanCancerGeneSignatureFiles}
\alias{processPanCancerGeneSignatureFiles}
\title{Read pan-cancer gene signature files and merge them into a `data.table`}
\usage{
processPanCancerGeneSignatureFiles(
  filePath,
  mDataType = c("rna", "cnv", "mutation"),
  fileFormats = c("rds")
)
}
\arguments{
\item{filePath}{`character(1)` Path to top level signature directory. This
should contain a folder `mDataType`.}

\item{mDataType}{`character(1)` The molecular data type of the signatures.}
}
\value{
`data.table` The gene signatures properly formatted to a single 
  long format table.
}
\description{
Read pan-cancer gene signature files and merge them into a `data.table`
}
