% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGeneSignatureToDT.R
\name{readGeneSig}
\alias{readGeneSig}
\title{Read a gene signature from a binary file format}
\usage{
readGeneSig(path, reader = NA)
}
\arguments{
\item{path}{Path to read the file from. Uses regex to extract the file type
and use the correct reader.}

\item{reader}{An optional function to read files. Use this for unsupported
file formats. Please pass the function directly, not a as a character
string.}
}
\description{
Read a gene signature from a binary file format
}
\details{
Currently supports .rds (readRDS) or .qs (qread) files.
}
