% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGeneSignatureToDT.R
\name{readGeneSigsForPSet}
\alias{readGeneSigsForPSet}
\title{Takes in the path to a set of gene signature .rds files, selects those pathing `pSetPattern` and converts them}
\usage{
readGeneSigsForPSet(dataDir, pSetPattern, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{dataDir}{[`character`] vector specifying the path to the gene signature .rds files}

\item{pSetPattern}{[`character`] an identifier for the Pset you wish to read. In general this will be the
pSet indentifier (e.g., GDSC_v1). '.*' is automatically prepended and appended to the pattern, so adding
this is not necssary.}

\item{BPPARAM}{[`BiocParam`] object specifying what backend to use for parallelization. This defaults to your
registered back-end, retrieved via the `bpparam()` function.}
}
\value{
[`list`] of `PharmacoSig` objects for the selected pSet
}
\description{
Takes in the path to a set of gene signature .rds files, selects those pathing `pSetPattern` and converts them
}
