% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGeneSignatureFiles.R
\name{readSigToDT}
\alias{readSigToDT}
\title{Read a gene signature file, then convert it to a `data.table`}
\usage{
readSigToDT(filePath)
}
\arguments{
\item{filePath}{`character(1)` The path to the signature file.}
}
\value{
`data.table` The contents of the gene signature long format.
}
\description{
Read a gene signature file, then convert it to a `data.table`
}
