% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNonASCII.R
\name{removeNonASCII}
\alias{removeNonASCII}
\title{Remove all non-ascii characters from a data.frame}
\usage{
removeNonASCII(df)
}
\arguments{
\item{df}{[`data.frame`] to purge of non-ASCII character}
}
\description{
Remove all non-ascii characters from a data.frame
}
