% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripEnsemblVersions.R
\name{stripEnsemblVersion,PharmacoSet-method}
\alias{stripEnsemblVersion,PharmacoSet-method}
\title{Remove version numbers from}
\usage{
\S4method{stripEnsemblVersion}{PharmacoSet}(object)
}
\arguments{
\item{object}{\code{PharmacoSet}}
}
\value{
The \code{PharmacoSet} object with all Ensembl ids regexed to remove
the version number.
}
\description{
Remove version numbers from
}
\details{
More specifically, this function will regex the \code{gene_id} column and the
\code{rownames()} of \code{rowData()} for all \code{SummarizedExperiment}s in the
\verb{@molecularProfiles} slot.
}
