% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToCsv-methods.R
\name{writeToCsv,PharmacoSet-method}
\alias{writeToCsv,PharmacoSet-method}
\title{Convert all data in an `PharmacoSet` object to disk as .csv files, named per}
\usage{
\S4method{writeToCsv}{PharmacoSet}(object, filePath)
}
\arguments{
\item{object}{[`S4`] An object inheriting from the `CoreGx::CoreSet` class, such as a
`PharmacoSet`, `RadioSet`, `ToxicoSet` or `XevaSet`.}

\item{filePath}{[`character`] The path to which the .csv files will be written}
}
\description{
Convert all data in an `PharmacoSet` object to disk as .csv files, named per
}
\section{Note}{

  This function will create a directory with the same name as the `object`
  being written to .csv, unless one already exists. The `filePath` is automatically
  appended with the object name, therefore it is not necssary to specify the object
  name in `filePath`
}

