% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToParquet-methods.R
\name{writeToParquet,PharmacoSet-method}
\alias{writeToParquet,PharmacoSet-method}
\title{Convert all data in an `PharmacoSet` object to disk as .parquet files,
  named according to the nested location in the PharmacoSet}
\usage{
\S4method{writeToParquet}{PharmacoSet}(object, filePath, ...)
}
\arguments{
\item{object}{[`S4`] An object inheriting from the `CoreGx::CoreSet` class, such as a
`PharmacoSet`, `RadioSet`, `ToxicoSet` or `XevaSet`.}

\item{filePath}{[`character`] The path to which the .parquet files will be written}

\item{...}{`pairlist` Fall through arguments to all instances of 
`arrow::write_parquet`}
}
\description{
Convert all data in an `PharmacoSet` object to disk as .parquet files,
  named according to the nested location in the PharmacoSet
}
\section{Note}{

  This function will create a directory with the same name as the `object`
  being written to .parquet, unless one already exists. The 
  not necssary to specify the object name in `filePath`
}

