% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{writeToParquet}
\alias{writeToParquet}
\title{Generic for writing an R object to one or more csv files}
\usage{
writeToParquet(object, ...)
}
\arguments{
\item{object}{[`object`] An R object to dump to csv}

\item{...}{[`pairlist`] Allow definition of new arguments in setMethod}
}
\description{
Generic for writing an R object to one or more csv files
}
