% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zeroPad}
\alias{zeroPad}
\title{Prepend zeros to an integer until its string representation has `string_length`
  characters.}
\usage{
zeroPad(x, string_length)
}
\arguments{
\item{x}{`integer(1)` An integer to prepend zeros to so the returned string
length of `string_length`.}

\item{string_length}{`integer(1)` The desired length of the returned string.}
}
\description{
Prepend zeros to an integer until its string representation has `string_length`
  characters.
}
