% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToCsv-methods.R
\name{.writeAnnotationToTxt}
\alias{.writeAnnotationToTxt}
\title{Write the annotation listt in a PharmacoSet to a text file}
\usage{
.writeAnnotationToTxt(
  annotations,
  dsType,
  filePath,
  objectName,
  sepChar = "|||"
)
}
\arguments{
\item{annotations}{[`list`] Annotations as returned by the `annotaiton` function}

\item{objectName}{[`character`] The name of the PSet the annotations are from.}

\item{sepChar}{[`character`] Separator to use when pasting together multiple lines of a list item.}

\item{path}{[`character`] The path to save the output file to.}
}
\value{
Writes to disk, does not return.
}
\description{
Each list item is one line, lines in each item are separated with `sepChar` so they
 can be split into to their original stucture.
}
\keyword{internal}
