% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGeneSignatureFiles.R
\name{processPanCancerGeneSignatureFiles}
\alias{processPanCancerGeneSignatureFiles}
\title{Read pan-cancer gene signature files and merge them into a `data.table`}
\usage{
processPanCancerGeneSignatureFiles(
  filePath,
  mDataType = c("rna", "cnv", "mutation"),
  ...,
  fileFormats = c("rds", "qs")
)
}
\arguments{
\item{filePath}{`character(1)` Path to top level signature directory. This
should contain a folder `mDataType`.}

\item{mDataType}{`character(1)` The molecular data type of the signatures.}

\item{...}{Fall through arguments to [`BiocParllel::bplapply`].}

\item{fileFormats}{`character` File formats to search in `filePath` for,
the dot should be excluded (e.g., rds not .rds, etc)}
}
\value{
`data.table` The gene signatures properly formatted to a single
  long format table.
}
\description{
Read pan-cancer gene signature files and merge them into a `data.table`
}
